requirements.txt
pyTelegramBotAPI
google-genai>=1.0.0
aiohttp
md2tgmd
Pillow
python-dotenv